fixed progression crash

+++ game/player.h
@@ -2,5 +2,5 @@
 #define __PLAYER_H
 
 object_t*
-player_init(uint32_t id, uint16_t animId, int16_t x);
+player_init(uint16_t id, uint16_t animId, int16_t x);
 #endif
